requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-HTTPAuth==4.8.0
python-dotenv==1.0.1
email-validator==2.1.1
WTForms==3.0.1
WeasyPrint==60.2
xhtml2pdf==0.2.13
qrcode==7.4.2
Pillow==10.3.0